p1: config: disable p1 hardware keyboard
This driver always reports itself as present even when the keyboard
is not.  This causes the software keyboard to be disabled.  Almost no
one has this accessory anyway - so disable it until someone who
actually has a keyboard to test with can fix this broken driver.

OR

Fix it from userspace:
https://gerrit.omnirom.org/#/c/3011/

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: Id6608367910d51fe1a2176693cbe7b356e59b43d